scrapers/requirements.txt
attrs==25.3.0
certifi==2025.6.15
charset-normalizer==3.4.2
h11==0.16.0
idna==3.10
outcome==1.3.0.post0
packaging==25.0
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
requests==2.32.4
selenium==4.34.0
six==1.17.0
sniffio==1.3.1
sortedcontainers==2.4.0
trio==0.30.0
trio-websocket==0.12.2
typing_extensions==4.14.0
urllib3==2.4.0
webdriver-manager==4.0.2
websocket-client==1.8.0
wsproto==1.2.0
